brokerage/requirements.txt
Django==2.0.2
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.7.7
requests==2.18.4